OneFragmentからCustomAdapterを分離
