+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/ui/search/OneFragment.kt
@@ -1,23 +1,15 @@
-/*
- * Copyright © 2021 YUMEMI Inc. All rights reserved.
- */
 package jp.co.yumemi.android.code_check.ui.search
 
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.view.inputmethod.EditorInfo
-import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
-import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.ListAdapter
-import androidx.recyclerview.widget.RecyclerView
 import jp.co.yumemi.android.code_check.R
 import jp.co.yumemi.android.code_check.data.model.item
+import jp.co.yumemi.android.code_check.data.repository.GithubRepository
 import jp.co.yumemi.android.code_check.databinding.FragmentOneBinding
 
 class OneFragment : Fragment(R.layout.fragment_one) {
@@ -26,14 +18,10 @@ class OneFragment : Fragment(R.layout.fragment_one) {
         savedInstanceState: Bundle?,
     ) {
         super.onViewCreated(view, savedInstanceState)
-
         val _binding = FragmentOneBinding.bind(view)
-
-        val _viewModel = OneViewModel(context!!)
-
-        val _layoutManager = LinearLayoutManager(context!!)
-        val _dividerItemDecoration =
-            DividerItemDecoration(context!!, _layoutManager.orientation)
+        val _viewModel = OneViewModel(GithubRepository(requireContext()))
+        val _layoutManager = LinearLayoutManager(requireContext())
+        val _dividerItemDecoration = DividerItemDecoration(requireContext(), _layoutManager.orientation)
         val _adapter =
             CustomAdapter(
                 object : CustomAdapter.OnItemClickListener {
@@ -43,79 +31,27 @@ class OneFragment : Fragment(R.layout.fragment_one) {
                 },
             )
 
-        _binding.searchInputText
-            .setOnEditorActionListener { editText, action, _ ->
-                if (action == EditorInfo.IME_ACTION_SEARCH) {
-                    editText.text.toString().let {
-                        _viewModel.searchResults(it).apply {
-                            _adapter.submitList(this)
-                        }
+        _binding.searchInputText.setOnEditorActionListener { editText, action, _ ->
+            if (action == EditorInfo.IME_ACTION_SEARCH) {
+                editText.text.toString().let {
+                    _viewModel.searchResults(it).observe(viewLifecycleOwner) {
+                        _adapter.submitList(it)
                     }
-                    return@setOnEditorActionListener true
                 }
-                return@setOnEditorActionListener false
+                return@setOnEditorActionListener true
             }
+            false
+        }
 
-        _binding.recyclerView.also {
-            it.layoutManager = _layoutManager
-            it.addItemDecoration(_dividerItemDecoration)
-            it.adapter = _adapter
+        _binding.recyclerView.apply {
+            layoutManager = _layoutManager
+            addItemDecoration(_dividerItemDecoration)
+            adapter = _adapter
         }
     }
 
-    fun gotoRepositoryFragment(item: item) {
-        val _action =
-            OneFragmentDirections.actionRepositoriesFragmentToRepositoryFragment(item = item)
+    private fun gotoRepositoryFragment(item: item) {
+        val _action = OneFragmentDirections.actionRepositoriesFragmentToRepositoryFragment(item = item)
         findNavController().navigate(_action)
-    }
-}
-
-val diff_util =
-    object : DiffUtil.ItemCallback<item>() {
-        override fun areItemsTheSame(
-            oldItem: item,
-            newItem: item,
-        ): Boolean {
-            return oldItem.name == newItem.name
-        }
-
-        override fun areContentsTheSame(
-            oldItem: item,
-            newItem: item,
-        ): Boolean {
-            return oldItem == newItem
-        }
-    }
-
-class CustomAdapter(
-    private val itemClickListener: OnItemClickListener,
-) : ListAdapter<item, CustomAdapter.ViewHolder>(diff_util) {
-    class ViewHolder(view: View) : RecyclerView.ViewHolder(view)
-
-    interface OnItemClickListener {
-        fun itemClick(item: item)
-    }
-
-    override fun onCreateViewHolder(
-        parent: ViewGroup,
-        viewType: Int,
-    ): ViewHolder {
-        val _view =
-            LayoutInflater.from(parent.context)
-                .inflate(R.layout.layout_item, parent, false)
-        return ViewHolder(_view)
-    }
-
-    override fun onBindViewHolder(
-        holder: ViewHolder,
-        position: Int,
-    ) {
-        val _item = getItem(position)
-        (holder.itemView.findViewById<View>(R.id.repositoryNameView) as TextView).text =
-            _item.name
-
-        holder.itemView.setOnClickListener {
-            itemClickListener.itemClick(_item)
-        }
     }
 }
